# openapi.yaml
openapi: 3.1.0
info:
  title: AI Stock Analysis API
  description: A microservice for AI-powered stock and cryptocurrency analysis using yfinance and GPT-4o-mini
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.stockanalysis.com
    description: Production server
...
paths:
...
  /api/v1/stocks/{symbol}:
    get:
      tags:
        - stocks
      summary: Get Stock
      description: Get individual stock/crypto data with historical prices and technical indicators.
      operationId: get_stock
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
          description: Stock symbol (e.g., AAPL, BTC-USD, NPN.JO)
          example: AAPL
        - name: market
          in: query
          required: false
          schema:
            type: string
            default: US
            enum: [US, ZA, CRYPTO]
          description: Market type
          example: US
        - name: period
          in: query
          required: false
          schema:
            type: string
            default: 1mo
            enum: [1d, 5d, 1mo, 3mo, 6mo, 1y, 2y, 5y, 10y, ytd, max]
          description: Time period for historical data
          example: 1mo
        - name: interval
          in: query
          required: false
          schema:
            type: string
            default: 1d
            enum: [1m, 2m, 5m, 15m, 30m, 60m, 90m, 1h, 1d, 5d, 1wk, 1mo, 3mo]
          description: Data interval
          example: 1d
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockData'
        '404':
          description: Stock Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/stocks/search:
    get:
      tags:
        - stocks
      summary: Search Stocks
      description: Search for stocks/crypto by name or ticker symbol.
      operationId: search_stocks
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
            minLength: 1
          description: Search query (company name or ticker)
          example: apple
        - name: market
          in: query
          required: false
          schema:
            type: string
            default: ALL
            enum: [US, ZA, CRYPTO, ALL]
          description: Market filter
          example: ALL
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
          description: Maximum number of results
          example: 10
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockSearchResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/stocks/batch:
    post:
      tags:
        - stocks
      summary: Get Batch Stocks
      description: Fetch multiple stocks in one request.
      operationId: get_batch_stocks
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchStockRequest'
            example:
              symbols: ["AAPL", "MSFT", "GOOGL"]
              period: "1mo"
              interval: "1d"
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchStockResponse'
        '404':
          description: No Stocks Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    StockData:
      type: object
      required:
        - symbol
        - market
        - company_name
        - current_price
        - currency
        - change_percent
        - volume
        - historical_data
        - technical_indicators
      properties:
        symbol:
          type: string
          description: Stock ticker symbol
          example: AAPL
        market:
          type: string
          enum: [US, ZA, CRYPTO]
          description: Market type
        company_name:
          type: string
          description: Company name
          example: Apple Inc.
        current_price:
          type: number
          description: Current stock price
          example: 150.25
        currency:
          type: string
          description: Currency code
          example: USD
        change_percent:
          type: number
          description: Daily change percentage
          example: 2.5
        volume:
          type: integer
          description: Trading volume
          example: 75000000
        market_cap:
          type: number
          nullable: true
          description: Market capitalization
        pe_ratio:
          type: number
          nullable: true
          description: Price-to-earnings ratio
        dividend_yield:
          type: number
          nullable: true
          description: Dividend yield percentage
        historical_data:
          type: array
          items:
            $ref: '#/components/schemas/HistoricalDataPoint'
        technical_indicators:
          $ref: '#/components/schemas/TechnicalIndicators'
        fifty_two_week_high:
          type: number
          nullable: true
          description: 52-week high price
        fifty_two_week_low:
          type: number
          nullable: true
          description: 52-week low price
        average_volume:
          type: integer
          nullable: true
          description: Average trading volume
        price_changes:
          type: object
          description: Price changes over different periods
    HistoricalDataPoint:
      type: object
      required:
        - timestamp
        - open
        - high
        - low
        - close
        - volume
      properties:
        timestamp:
          type: string
          format: date-time
        open:
          type: number
        high:
          type: number
        low:
          type: number
        close:
          type: number
        volume:
          type: integer
    TechnicalIndicators:
      type: object
      properties:
        rsi:
          type: number
          nullable: true
          description: 14-period RSI
        sma_20:
          type: number
          nullable: true
          description: 20-day Simple Moving Average
        sma_50:
          type: number
          nullable: true
          description: 50-day Simple Moving Average
        sma_200:
          type: number
          nullable: true
          description: 200-day Simple Moving Average
        macd:
          type: number
          nullable: true
          description: MACD line
        macd_signal:
          type: number
          nullable: true
          description: MACD signal line
        macd_histogram:
          type: number
          nullable: true
          description: MACD histogram
    StockSearchResponse:
      type: object
      required:
        - results
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/StockSearchResult'
    StockSearchResult:
      type: object
      required:
        - symbol
        - name
        - market
      properties:
        symbol:
          type: string
        name:
          type: string
        market:
          type: string
          enum: [US, ZA, CRYPTO]
        sector:
          type: string
          nullable: true
    BatchStockRequest:
      type: object
      required:
        - symbols
      properties:
        symbols:
          type: array
          items:
            type: string
          maxItems: 10
          minItems: 1
        period:
          type: string
          default: 1mo
        interval:
          type: string
          default: 1d
    BatchStockResponse:
      type: object
      required:
        - stocks
      properties:
        stocks:
          type: array
          items:
            $ref: '#/components/schemas/StockData'
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: object
          properties:
            error:
              type: object
              properties:
                code:
                  type: string
                message:
                  type: string
                details:
                  type: object
        timestamp:
          type: string
          format: date-time
    HTTPValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
    ValidationError:
      type: object
      required:
        - loc
        - msg
        - type
      properties:
        loc:
          type: array
          items:
            anyOf:
              - type: string
              - type: integer
        msg:
          type: string
        type:
          type: string
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Optional API key for rate limiting and access control
security:
  - ApiKeyAuth: []
  - {}  # Allow access without API key
...